Update PaymentActionsValidator.php
An error was encountered when attempting to invoice an order paid through Affirm. The error was encountered when attempting to log the error message, specifically when the $validationResult variable was set to false, indicating that the response could not be validated. Since the translation has already been rendered, the $errorMessages[0] variable is of type string and the \Magento\Framework\Phrase::render() function cannot be used.